# src/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ANON - Anonymous Employee Platform API",
        "description": "API documentation for the Anonymous Employee Platform. This platform allows employees to post anonymous messages and interact with each other.",
        "contact": {
            "name": "API Support",
            "email": "support@company.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.company.com",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/broadcasting/auth": {
            "post": {
                "tags": [
                    "Broadcasting"
                ],
                "summary": "Authenticate WebSocket channel subscription",
                "description": "Authenticates a user for WebSocket channel subscription. This endpoint is used by Laravel Echo to authorize private channel subscriptions.",
                "operationId": "ce6fbccfa2696401d7f6dda2307b0266",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "channel_name",
                                    "socket_id"
                                ],
                                "properties": {
                                    "channel_name": {
                                        "description": "The name of the channel to subscribe to",
                                        "type": "string",
                                        "example": "private-chat.14"
                                    },
                                    "socket_id": {
                                        "description": "The socket ID from the WebSocket connection",
                                        "type": "string",
                                        "example": "123456.789012"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "auth": {
                                            "description": "The authentication signature",
                                            "type": "string",
                                            "example": "DRcpFZv3w5a4:123456.789012:abcdef1234567890"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - missing parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Missing channel_name or socket_id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - user not authorized for this channel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}